+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: rhutchin <rhutchin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/20 09:16:31 by rhutchin          #+#    #+#             */
-/*   Updated: 2019/07/23 15:03:47 by rhutchin         ###   ########.fr       */
+/*   Updated: 2019/07/24 08:19:28 by rhutchin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef struct		s_list
 
 int					ft_atoi(const char *str);
 char				*ft_itoa_base(long value, int base);
-size_t				ft_intlen(long int num);
+int					ft_intlen(int num);
 
 void				ft_bzero(void *str, size_t n);
